refactor: add questions factories
1. Create questions factories for isolating questions generation from the main process.
2. Change props metadata structure for taking description to improve questions format
