+++ src/questions/index.ts
@@ -1,0 +1,2 @@
+export { default as forSections } from './forSections'
+export { default as forProps } from './forProps'
