+++ teste.md
@@ -1,0 +1,42 @@
+
+# create-md
+
+> Create docs with a simple command
+
+![GitHub last commit](https://img.shields.io/github/last-commit/lucas-lm/create-md)
+![license](https://img.shields.io/github/license/lucas-lm/create-md)
+![GitHub repo size](https://img.shields.io/github/repo-size/lucas-lm/create-md)
+
+## About
+
+
+
+## Prerequisites
+
+- node 
+- npm 
+
+## Contributing
+
+Contributions, issues and feature requests are welcome!
+
+- ⭐️ Star the project
+- 🐛 Find and report issues
+- 📥 Submit PRs to help solve issues or add features
+
+Feel free to check [issues page](https://github.com//create-md/issues). You can also take a look at the contributing guide.
+
+## Author
+
+**Lucas Miranda** lucas-m@outlook.com
+* Website: https://lucas-lm.github.io
+
+* GitHub: [@lucas-lm](https://github.com/lucas-lm)
+
+
+
+
+## 📝 License
+
+Copyright © 2020 Lucas Miranda
+This project is [MIT](license) licensed.
